update to new plugin structure
